update movies, added owner field to movies, update jwt-auth models

+++ comments/views.py
@@ -1,19 +1,23 @@
+#from django.core.exceptions import PermissionDenied
+#from rest_framework import permissions
 from comments.models import Comment
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
 from rest_framework.exceptions import NotFound
-#from rest_framework.permissions import IsAuthenticated
-#from rest_framework.exceptions import PermissionDenied
+from rest_framework.permissions import IsAuthenticated
+from rest_framework.exceptions import PermissionDenied
 
 
 from .serializers.common import CommentSerializer
-#from .models import Comment
+from .models import Comment
 
 
 class CommentListView(APIView):
+    permission_classes = (IsAuthenticated, )
 
     def post(self, request):
+        request.data["owner"] = request.user.id
         comment_to_create = CommentSerializer(data=request.data)
         if comment_to_create.is_valid():
             comment_to_create.save()
@@ -23,10 +27,12 @@ class CommentListView(APIView):
 
 class CommentDetailView(APIView):
 
-    def delete(self, _request, pk):
+    def delete(self, request, pk):
         try:
             comment_to_delete = Comment.objects.get(pk=pk)
         except Comment.DoesNotExist:
             raise NotFound()
+        if comment_to_delete.owner != request.user:
+            raise PermissionDenied()
         comment_to_delete.delete()  # delete from table
         return Response(status=status.HTTP_204_NO_CONTENT)
